debezium/dbz#453 Add GENERIC group and use fieldsByGroup() as source of truth for group assignment
